fix live radio page tests
